feat: support augmenting trace context properties
Add a property to SpanOptions to support optionally augmenting log
events with additional properties to represent the trace context. This
is to support the case where the destination that logs will be shipped
to expects different property names or differently formatted values to
represent traces than what is used natively by the application.